requirements.txt
Django==1.6.5
Fabric==1.9.0
Pillow==2.5.0
argparse==1.2.1
boto==2.31.1
distribute==0.6.24
django-configurations==0.8
django-registration==1.0
ecdsa==0.11
paramiko==1.14.0
postgres==2.1.2
psycopg2==2.5.3
pycrypto==2.6.1
requirements==0.1
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
